Admin notes
From session booker, single-click a request to open the request viewer

The request viewer will have a texture for admin notes.

When the texture for admin notes is changes and focus leaves, the admin notes are updated on the request.

Requests cards will all display the admin notes, except for in the Driver Link and Musician Link pages.